requirements.txt
tensorflow==2.15.0
keras==2.15.0
deepface==0.0.91
opencv-python-headless==4.10.0.84
pandas==2.2.3
scikit-learn==1.5.2
matplotlib==3.9.2
numpy==1.24.4
seaborn==0.13.2
streamlit==1.38.0
pillow==10.4.0
mtcnn==0.1.1